Fixed parser initialiser and parser destroyer
The parser initiator was not able to detect, whether a return of NULL meant an error (like failed malloc) or whether we're at the end of the token. and generally would
not have cared if there was a failed malloc. Now malloc checks are in place and a check variable is in t_command which is set to one upon an error.

There is still the issue of free_ing the whole list of commands, if there is an error. At the moment only the erroneous link is removed and the parsing stops.