Parser opens files
The parser opens the fds for all in- and outfile in the commandline, as bash creates or attempts to open all files named.
So, in order to avoid earlier named files to be overwritten, we need to open/create and check them as they come through. The last mentioned of its type
is what is used as the source or destination.

Still plenty of inconsistencies among the files for redirections